# design/bus-eta-openapi.yaml
openapi: 3.0.3
info:
  title: Hong Kong Transportation Aggregation API
  version: 1.0.0
  description: |
    Aggregated bus-ETA + weather + traffic-incident API for the Hong Kong
    transportation mobile client.

    ## Backend service
    A FastAPI service (package `transportation-api`) fans out, concurrently, to
    four public Hong Kong open-data feeds, normalises them into the canonical
    models below, and serves them through the endpoints in this document:

    | Concern    | Provider / feed                          | Client module          |
    |------------|------------------------------------------|------------------------|
    | Bus ETA    | KMB `data.etabus.gov.hk` stop-eta        | `src/clients/kmb.py`   |
    | Bus ETA    | Citybus/NWFB `rt.data.gov.hk` eta        | `src/clients/citybus.py` |
    | Weather    | HKO OpenData `weatherAPI`                | `src/clients/hko.py`   |
    | Incidents  | HK Transport Department traffic-news XML | `src/clients/td.py`    |

    Operator/stop-id namespaces are disjoint and decide routing:
      * KMB      — 16-char hex ids (e.g. `946C74E30100FE80`)
      * Citybus  — 6-digit numeric ids (e.g. `001027`)

    ## Canonical base path
    Every versioned endpoint in this contract lives under `/api/v1`. This is the
    base path the mobile frontend already consumes (`GET /api/v1/eta`,
    `GET /api/v1/search`).

    ### Known implementation deviations (reconcile on backend)
...

...

    ## Wire format conventions
...
        the others.
...
servers:
...
paths:
...
  /api/v1/eta:
    get:
      operationId: getEtaAggregate
      tags: [Aggregation]
      summary: >-
        Combined bus ETA + weather + traffic incidents for a route & stop
        (PRIMARY aggregation endpoint used by the mobile frontend)
      description: |
        Returns a single object containing the upcoming arrivals for a
        `(route, stop)` pair, the current HKO weather + active warnings, and
        any Transport Department traffic incidents relevant to that
        route/stop (each tagged with a server-computed `relevance` and sorted
        high-first). This is the endpoint the mobile "results" screen calls.
      parameters:
        - $ref: "#/components/parameters/RouteQuery"
        - $ref: "#/components/parameters/StopQuery"
        - $ref: "#/components/parameters/LangQuery"
        - $ref: "#/components/parameters/IncludeWeatherQuery"
        - $ref: "#/components/parameters/IncludeIncidentsQuery"
        - $ref: "#/components/parameters/DegradeQuery"
      responses:
        "200":
          description: >-
            Aggregated payload. `degraded` is true when a secondary provider
            failed and was skipped.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EtaAggregate"
              examples:
                ok:
                  $ref: "#/components/examples/EtaAggregateOk"
        "404":
          description: No operator returned an ETA for this route+stop.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: RESOURCE_NOT_FOUND
                message: "No ETA data for route 1 at stop 946C74E30100FE80."
                detail: "No operator (KMB / Citybus / NWFB) returned an ETA for this route+stop."
        "422":
          description: Missing required query parameter (FastAPI validation).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: >-
            Upstream provider failed and degradation is disabled
            (`degrade=false`), or an unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/weather:
    get:
      operationId: getWeather
      tags: [Weather]
      summary: Current Hong Kong weather, warnings, and optional forecast
      description: |
        Current conditions + active HKO warnings from `rhrread`/`warnsum`.
        Attach `include_forecast=true` for the 9-day forecast. Fail-soft: a
        missing core feed surfaces as `500 UPSTREAM_ERROR`.
      parameters:
        - $ref: "#/components/parameters/LangQuery"
        - name: include_forecast
          in: query
          required: false
          description: Attach the 9-day forecast to the response.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Current weather payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Weather"
        "500":
          description: HKO feed unreachable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/weather/warnings:
    get:
      operationId: getWeatherWarnings
      tags: [Weather]
      summary: Active Hong Kong weather warnings only
      parameters:
        - $ref: "#/components/parameters/LangQuery"
      responses:
        "200":
          description: Active warnings wrapper.
          content:
            application/json:
              schema:
                type: object
                required: [warnings]
                properties:
                  warnings:
                    type: array
                    items:
                      $ref: "#/components/schemas/WeatherWarning"
              examples:
                none:
                  value:
                    warnings: []
                one:
                  value:
                    warnings:
                      - code: WHOT
                        title:
                          en: Very Hot Weather Warning
                          tc: 酷熱天氣警告
                          sc: 酷热天气警告
                        severity: warning
                        contents: 酷熱天氣警告現正生效。
                        issue_time: "2026-07-10T11:30:00+08:00"
...
  /eta:
    get:
      operationId: getEtaLegacy
      deprecated: true
      tags: [Aggregation]
      summary: "[LEGACY] Flat bus ETA + alerts payload (do not use for new clients)"
      description: |
        Deprecated flat shape consumed by the legacy React Native client.
        New clients MUST use `GET /api/v1/eta`. Retained for backward
        compatibility only.
      parameters:
        - name: route
          in: query
          required: true
          schema:
            type: string
          description: Bus route number, e.g. '1'.
        - name: stop_id
          in: query
          required: true
          schema:
            type: string
          description: Bus stop id, e.g. '946C74E30100FE80'.
      responses:
        "200":
          description: Legacy flattened payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EtaResponse"
components:
  parameters:
    LangQuery:
      name: lang
      in: query
      required: false
      description: Language for text fields. One of en / tc / sc.
      schema:
        type: string
        enum: [en, tc, sc]
        default: tc
    RouteQuery:
      name: route
      in: query
      required: true
      description: Bus route number, e.g. '1'.
      schema:
        type: string
      example: "1"
    StopQuery:
      name: stop
      in: query
      required: true
      description: Bus stop id. KMB = 16-char hex; Citybus/NWFB = 6-digit numeric.
      schema:
        type: string
      example: "946C74E30100FE80"
    IncludeWeatherQuery:
      name: include_weather
      in: query
      required: false
      description: Include the Weather block.
      schema:
        type: boolean
        default: true
    IncludeIncidentsQuery:
      name: include_incidents
      in: query
      required: false
      description: Include the Incidents block.
      schema:
        type: boolean
        default: true
    DegradeQuery:
      name: degrade
      in: query
      required: false
      description: >-
        Fail-soft: on partial upstream failure return partial data
        (degraded=true) instead of 500.
      schema:
        type: boolean
        default: true
  examples:
    EtaAggregateOk:
      summary: Route 1 at a KMB stop, with weather + one incident
      value:
        query:
          route: "1"
          stop_id: "946C74E30100FE80"
          operator: KMB
          lang: tc
        etas:
          - co: KMB
            route: "1"
            direction: O
            service_type: 1
            seq: 8
            dest:
              en: STAR FERRY
              tc: 尖沙咀碼頭
              sc: 尖沙咀码头
            eta_seq: 1
            eta: "2026-07-12T08:42:00Z"
            minutes_remaining: 4
            remark:
              tc: 預定班次
            data_timestamp: "2026-07-12T08:37:55Z"
          - co: KMB
            route: "1"
            direction: O
            service_type: 1
            seq: 8
            dest:
              en: STAR FERRY
              tc: 尖沙咀碼頭
              sc: 尖沙咀码头
            eta_seq: 2
            eta: "2026-07-12T08:54:00Z"
            minutes_remaining: 16
            remark: null
            data_timestamp: "2026-07-12T08:37:55Z"
        weather:
          temperature:
            place: Hong Kong Observatory
            value: 28
            unit: C
          description: 天色良好
          humidity:
            value: 84
            unit: percent
          icon: [63]
          update_time: "2026-07-12T08:30:00Z"
          warnings:
            - code: WRAINA
              title:
                en: Amber Rainstorm Warning
                tc: 黃色暴雨警告信號
                sc: 黄色暴雨警告信号
              severity: amber
              contents: 黃色暴雨警告信號現正生效。
              issue_time: "2026-07-12T07:15:00Z"
          forecast: null
        incidents:
          - id: TD20260712-00123
            heading:
              tc: 因交通意外道路封閉
            detail:
              tc: 觀塘道交通意外，部分行車線封閉
            location:
              tc: 觀塘道
            district:
              tc: 觀塘
            status:
              tc: 生效中
            announcement_date: "2026-07-12 08:30"
            relevance: low
            content:
              tc: 觀塘道因交通意外部分行車線封閫。
            source_id: "141824"
            geo:
              lat: 22.3119
              lon: 114.2257
        query_time: "2026-07-12T08:38:00Z"
        degraded: false
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
        detail:
          type: string
          nullable: true
    MultilingualText:
      type: object
      description: Human-readable text in up to three languages.
      properties:
        en:
          type: string
          nullable: true
        tc:
          type: string
          nullable: true
        sc:
          type: string
          nullable: true
    GeoPoint:
      type: object
      required: [lat, lon]
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
        lon:
          type: number
          format: double
          minimum: -180
          maximum: 180
...
    ETA:
      type: object
      required: [co, route, direction, seq, dest, eta_seq]
      description: |
        Predicted arrival of a route at a stop. NOTE: the `status`
        (`live`/`scheduled`) field used by clients is DERIVED client-side and
        is intentionally not emitted by the backend.
      properties:
        co:
          type: string
          example: KMB
        route:
          type: string
          example: "1"
        direction:
          type: string
          description: "O = outbound, I = inbound."
          example: "O"
        service_type:
          type: integer
          default: 1
        seq:
          type: integer
          description: Stop sequence number on the route.
        dest:
          $ref: "#/components/schemas/MultilingualText"
        eta_seq:
          type: integer
          description: Ordering index among upcoming arrivals.
        eta:
          type: string
          format: date-time
          nullable: true
          description: Predicted arrival timestamp (UTC). null = no prediction.
        minutes_remaining:
          type: integer
          minimum: 0
          nullable: true
          description: Minutes until arrival, computed server-side (>=0). null when eta is null.
        remark:
          $ref: "#/components/schemas/MultilingualText"
          nullable: true
          description: Human remark, e.g. 'Scheduled', 'KMB staff on board'.
        data_timestamp:
          type: string
          format: date-time
          nullable: true
          description: When the operator last published this prediction.
    WeatherWarning:
      type: object
      required: [code, title, severity]
      properties:
        code:
          type: string
          description: HKO warning code, e.g. WRAINA.
          example: WRAINA
        title:
          $ref: "#/components/schemas/MultilingualText"
        severity:
          type: string
          enum: [none, amber, red, black, warning]
          example: amber
        contents:
          type: string
          nullable: true
          description: Full warning text issued by HKO.
        issue_time:
          type: string
          format: date-time
          nullable: true
    ForecastDay:
      type: object
      description: One day from the HKO 9-day forecast (optional enrichment).
      properties:
        date:
          type: string
          nullable: true
          description: Forecast date (YYYYMMDD).
        week:
          type: string
          nullable: true
          description: Day of week, localised.
        weather:
          type: string
          nullable: true
        max_temp:
          type: number
          nullable: true
        min_temp:
          type: number
          nullable: true
    Weather:
      type: object
      properties:
        temperature:
          type: object
          nullable: true
          properties:
            place:
              type: string
            value:
              type: number
            unit:
              type: string
          description: Current air temperature.
        description:
          type: string
          nullable: true
          description: >-
            Human-readable conditions resolved to `lang` (derived from HKO icon
            codes), e.g. 'Showers' / '多雲' / '多云'.
        humidity:
          type: object
          nullable: true
          properties:
            value:
              type: number
            unit:
              type: string
          description: Relative humidity.
        icon:
          type: array
          items:
            type: integer
          description: HKO weather icon codes.
        update_time:
          type: string
          format: date-time
          nullable: true
          description: Feed update time.
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/WeatherWarning"
          description: Active warnings.
        forecast:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ForecastDay"
          description: Short forecast (optional).
    Incident:
      type: object
      required: [id, heading, location, status]
      properties:
        id:
          type: string
          description: TD incident number.
          example: "TD20260710-00123"
        heading:
          $ref: "#/components/schemas/MultilingualText"
        detail:
          $ref: "#/components/schemas/MultilingualText"
          nullable: true
        location:
          $ref: "#/components/schemas/MultilingualText"
        district:
          $ref: "#/components/schemas/MultilingualText"
          nullable: true
        direction:
          $ref: "#/components/schemas/MultilingualText"
          nullable: true
        road_type:
          $ref: "#/components/schemas/MultilingualText"
          nullable: true
        near_landmark:
          $ref: "#/components/schemas/MultilingualText"
          nullable: true
        status:
          $ref: "#/components/schemas/MultilingualText"
          description: INCIDENT_STATUS_* textual value, e.g. ACTIVE / 生效中.
        announcement_date:
          type: string
          nullable: true
          description: e.g. '2026-07-10 08:30'.
          example: "2026-07-10 08:30"
        relevance:
          type: string
          nullable: true
          enum: [high, medium, low, none]
          description: >-
            Server-computed relevance to the queried route/stop; set only on
            combined endpoints.
        content:
          $ref: "#/components/schemas/MultilingualText"
          nullable: true
          description: Full incident narrative / impact text (TD CONTENT_* field).
        source_id:
          type: string
          nullable: true
          description: Numeric message id assigned by the TD feed (element <ID>).
        geo:
          $ref: "#/components/schemas/GeoPoint"
          nullable: true
          description: WGS84 coordinate when the feed provides LATITUDE/LONGITUDE.
    EtaQuery:
      type: object
      required: [route, stop_id, lang]
      properties:
        route:
          type: string
          description: Route number that was requested.
        stop_id:
          type: string
          description: Stop id that was requested.
        operator:
          type: string
          nullable: true
          description: Resolved operator (KMB / CTB / NWFB), or null if unknown.
        lang:
          type: string
          default: tc
          description: Language used for the response text.
    EtaAggregate:
      type: object
      required: [query, etas, query_time, degraded]
      description: >-
        PRIMARY aggregation response for GET /api/v1/eta. One object the
        mobile results screen renders.
      properties:
        query:
          $ref: "#/components/schemas/EtaQuery"
        etas:
          type: array
          items:
            $ref: "#/components/schemas/ETA"
          description: Upcoming arrivals for the route at the stop.
        weather:
          $ref: "#/components/schemas/Weather"
          nullable: true
          description: Current HKO weather + active warnings (null if degraded/omitted).
        incidents:
          type: array
          items:
            $ref: "#/components/schemas/Incident"
          description: >-
            Traffic incidents relevant to the route/stop, each tagged with a
            server-computed `relevance` (high/medium/low), sorted high-first.
        query_time:
          type: string
          format: date-time
          description: Server time the response was assembled (UTC).
        degraded:
          type: boolean
          default: false
          description: >-
            True when one or more providers failed and were skipped (partial
            data returned). Mirrors the fail-soft aggregation policy.
...
    EtaItem:
      type: object
      required: [route, dest]
      properties:
        route:
          type: string
        dest:
          type: string
        minutes_remaining:
          type: integer
          nullable: true
        eta_time:
          type: string
          nullable: true
        remark:
          type: string
    EtaResponse:
      type: object
      required: [query_time, bus_eta]
      properties:
        query_time:
          type: string
          format: date-time
        bus_eta:
          type: array
          items:
            $ref: "#/components/schemas/EtaItem"
        weather_warnings:
          type: array
          items:
            type: string
        traffic_incidents:
          type: array
          items:
            type: object
            additionalProperties: true